feat: ability to group and filter code blocks
With long markdown files it is not rare to have 4-5 or more code blocks,
but for different reasons under different sections of the document.

With this feature, we can specify key-value pairs on code blocks, and
then filter them using the `--group` cli flag.

```
❯ ./target/release/eval-md bash a.md
This is group A
This is group B
This one does not belon anywhere

❯ ./target/release/eval-md bash a.md --group=a
This is group A

❯ ./target/release/eval-md bash a.md --group=b
This is group B

❯ ./target/release/eval-md bash a.md --group=
This one does not belon anywhere
```

Details in the `README.md` file.

Signed-off-by: Efertone <efertone@pm.me>